+++ mi-moneda/src/main/java/co/uniquindio/edu/mi_moneda/model/Beneficio.java
@@ -18,7 +18,7 @@ public class Beneficio {
 
     private String nombre;
     private String descripcion;
-    private int costePuntos;
+    private double costePuntos;
     private TipoBeneficio tipo;
     private double valor;
 }
